fix(mobile): stop search indicator CLS and group preferences
Move the Agents search spinner into the search input. The separate
"Searching…" row appeared below the input and pushed the list down.
The spinner now swaps with the search icon in a fixed-size slot.

Move the Appearance control and the Notifications row from the Profile
screen into the Preferences screen.
