+++ apps/mobile/src/components/profile-screen.tsx
@@ -1,9 +1,8 @@
-/* eslint-disable max-lines -- The profile screen composes Credits, Agents, Reviews, Organization, Linked accounts, Notifications, Restore Purchases, and Actions; each section is a small rendered surface that mirrors the shared ConfigureRow/Text-header pattern. Splitting would re-encode the same hooks. */
+/* eslint-disable max-lines -- The profile screen composes Credits, Agents, Reviews, Organization, Linked accounts, App, Restore Purchases, and Actions; each section is a small rendered surface that mirrors the shared ConfigureRow/Text-header pattern. Splitting would re-encode the same hooks. */
 import { useMutation, useQuery } from '@tanstack/react-query';
 import * as Application from 'expo-application';
 import { type Href, useRouter } from 'expo-router';
 import {
-  Bell,
   Building2,
   GitMerge,
   GitPullRequest,
@@ -26,7 +25,6 @@ import { QueryError } from '@/components/query-error';
 import { ScreenHeader } from '@/components/screen-header';
 import { TabScreenScrollView } from '@/components/tab-screen';
 import { ConfigureRow } from '@/components/ui/configure-row';
-import { SegmentedControl } from '@/components/ui/segmented-control';
 import { Skeleton } from '@/components/ui/skeleton';
 import { Text } from '@/components/ui/text';
 import { FEATURE_FLAG_PR_REVIEW, useFeatureFlag } from '@/lib/analytics/posthog';
@@ -34,11 +32,6 @@ import { useAuth } from '@/lib/auth/auth-context';
 import { showFeedbackPrompt } from '@/lib/feedback';
 import { useCurrentUserId } from '@/lib/hooks/use-current-user-id';
 import { useThemeColors } from '@/lib/hooks/use-theme-colors';
-import {
-  setThemePreference,
-  type ThemePreference,
-  useThemePreference,
-} from '@/lib/hooks/use-theme-preference';
 import { useOrganization } from '@/lib/organization-context';
 import {
   getCodeReviewerProfilePath,
@@ -57,7 +50,6 @@ export function ProfileScreen() {
   const router = useRouter();
   const trpc = useTRPC();
   const colors = useThemeColors();
-  const { preference: themePreference } = useThemePreference();
   const { organizationId, isLoaded: organizationContextLoaded } = useOrganization();
   const isAuthenticated = token != null;
   const prReviewEnabled = useFeatureFlag(FEATURE_FLAG_PR_REVIEW, true);
@@ -275,23 +267,6 @@ export function ProfileScreen() {
           </View>
         )}
 
-        {/* Appearance */}
-        <View className="mt-6 gap-3">
-          <Text variant="small" className="uppercase tracking-wide text-muted-foreground">
-            Appearance
-          </Text>
-          <SegmentedControl<ThemePreference>
-            accessibilityLabel="Appearance"
-            options={[
-              { value: 'system', label: 'System' },
-              { value: 'light', label: 'Light' },
-              { value: 'dark', label: 'Dark' },
-            ]}
-            value={themePreference}
-            onChange={setThemePreference}
-          />
-        </View>
-
         {/* App */}
         <View className="mt-6 gap-3">
           <Text variant="small" className="uppercase tracking-wide text-muted-foreground">
@@ -300,28 +275,11 @@ export function ProfileScreen() {
           <ConfigureRow
             icon={SlidersHorizontal}
             title="Preferences"
-            subtitle="Thinking and screen behavior"
+            subtitle="Appearance, notifications, thinking, and screen behavior"
             className="rounded-lg bg-secondary px-3"
             last
             onPress={() => {
               router.push('/(app)/(tabs)/(3_profile)/preferences' as Href);
-            }}
-          />
-        </View>
-
-        {/* Notifications */}
-        <View className="mt-6 gap-3">
-          <Text variant="small" className="uppercase tracking-wide text-muted-foreground">
-            Notifications
-          </Text>
-          <ConfigureRow
-            icon={Bell}
-            title="Notifications"
-            subtitle="Push preferences"
-            className="rounded-lg bg-secondary px-3"
-            last
-            onPress={() => {
-              router.push('/(app)/(tabs)/(3_profile)/notifications' as Href);
             }}
           />
         </View>
